AddSec/requirements.txt
ruamel.yaml.clib==0.2.6
Flask==2.0.1
Werkzeug==2.0.1
keyring==22.2.0
JPype1==1.3.0
Flask_Cors==3.0.10
backports.statistics==0.1.0
configobj==5.0.6
Cython==0.29.24
ipaddr==2.2.0
mypy_extensions==0.4.3
numpy==1.21.1
ordereddict==1.1
typing_extensions==3.10.0.0
wincertstore==0.2.1
